Code refactoring and added timed hosted service

+++ WebApplication/DTO/ShipCaptainDTO.cs
@@ -4,9 +4,9 @@ namespace WebShipPort.DTO
 {
     public class ShipCaptainDTO : WorkerDTO
     {
-        public int SailingHoursTotal { get; set; }
+        public double SailingHoursTotal { get; set; }
 
-        public int SailingHoursAsCaptain { get; set; }
+        public double SailingHoursAsCaptain { get; set; }
 
         public ShipCaptainDTO() { }
